requirements.txt
asgiref==3.8.1
Django==4.2.8
django-cors-headers==4.4.0
djangorestframework==3.15.2
fingerprint==0.1.6
imageio==2.34.1
lazy_loader==0.4
networkx==3.3
numpy==2.0.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.3.0
scikit-image==0.24.0
scipy==1.13.1
sqlparse==0.5.0
tifffile==2024.6.18
tzdata==2024.1
